feat : fileDetail Long -> int로 수정, 파일 저장위치 공유폴더로 수정

+++ src/main/java/kr/or/ddit/util/file/service/FileDetailVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class FileDetailVO {
-    private Long fileId;             // UUID or 시퀀스
+    private int fileId;             // UUID or 시퀀스
     private Long fileGroupId;        // FK
     private String fileOrgName;        // 원본 파일명
     private String fileSaveName;       // 저장된 파일명 (uuid_원본)
